Remove LOG_DIR - it's no longer needed
Originally, the logging setup was writing to a file as well as the
console. The idea was to have a log file locally to veryify that
logging statments actually contained actionable information. The
reality that the colorized output to the console, created by
structlog, was a lot more useful, so the writing to a local log file
was dropped. This change to the settings takes care of a loose end
that was missed when the logging configuration was updated.

To keep things aligned with the 12-factor app philosophy,
writing to the log file was dropped since it was not reallly being
used.